# fix ts perf regression