--- a/shared/constants/types/fs.tsx
+++ b/shared/constants/types/fs.tsx
@@ -2,9 +2,6 @@ import * as RPCTypes from './rpc-gen'
 import * as ChatTypes from './chat2'
 import * as Devices from './devices'
 import * as TeamsTypes from './teams'
-// TODO importing FsGen causes an import loop
-import * as FsGen from '../../actions/fs-gen'
-import * as EngineGen from '../../actions/engine-gen-gen'
 import {isWindows} from '../platform'
 import {memoize} from '../../util/memoize'
 // lets not create cycles in flow, lets discuss how to fix this
@@ -27,8 +24,8 @@ export enum ProgressType {
 export type FsError = Readonly<{
   time: number
   errorMessage: string
-  erroredAction: FsGen.Actions | EngineGen.Actions
-  retriableAction?: FsGen.Actions | EngineGen.Actions
+  erroredAction: any // FsGen.Actions | EngineGen.Actions // using this type in the actions itself causes an explosive loop
+  retriableAction?: any // FsGen.Actions | EngineGen.Actions
 }>
 
 export type Device = Readonly<{